requirement.txt
pyaml==23.9.7
munch==4.0.0
numpy==1.26.1
torch==1.12.0
torchvision==0.13.0
torchaudio==0.12.0
librosa==0.10.1
git+https://github.com/resemble-ai/monotonic_align.git
SoundFile==0.12.1
pydub==0.25.1
matplotlib==3.8.1
phonemizer==3.2.1
attrdict==2.0.1
denoiser==0.1.5
mutagen==1.47.0
moviepy==1.0.3
pytube==15.0.0
ffmpeg==1.4
yt_dlp==2023.10.13
unidecode==1.3.7
fastapi==0.85.1
starlette==0.20.4
uvicorn==0.18.3
python-multipart==0.0.6